Rename const util Class
